ta lindo, customer name, criar e aceitar oferta detalhada, cod catch, validações, lancamentos

+++ Eice.Payment.API/DTO/LancamentoDto.cs
@@ -5,13 +5,13 @@ namespace Eice.Payment.API.DTO
 {
     public class LancamentoDto
     {
-        public int Saldo { get; set; }
+        public decimal Saldo { get; set; }
         public List<LancamentoItemDto> LancamentoItems { get; set; } = new List<LancamentoItemDto>();
     }
 
     public class LancamentoItemDto
     {
-        public int Quantidade { get; set; }
+        public decimal Quantidade { get; set; }
         public string Description { get; set; }
         public DateTime CreationTime { get; set; }
     }
