+++ Eice.Payment.API/DTO/OfertaDto.cs
@@ -8,8 +8,8 @@ namespace Eice.Payment.API.DTO
         public string Id { get; set; }
         public string CoinIdOffer { get; set; }
         public string CoinIdReceive { get; set; }
-        public int QuantityOffer { get; set; }
-        public int QuantityReceive { get; set; }
+        public decimal QuantityOffer { get; set; }
+        public decimal QuantityReceive { get; set; }
         public string Status { get; set; }
         public DateTime CreationTime { get; set; }
     }
